fix(lint): 린트 프로세스 변경
