+++ package.json
@@ -23,6 +23,7 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-storybook": "^0.6.12",
     "hygen": "^6.2.11",
     "jest": "^29.5.0",
@@ -35,8 +36,10 @@
     "typescript": "^5.0.4"
   },
   "scripts": {
+    "multitask": "turbo run lint test build",
     "test": "turbo run test",
-    "build": "turbo build --filter=!@danji/docs --filter=!@danji/rollup",
+    "lint": "turbo run lint",
+    "build": "turbo build --filter=!@danji/docs",
     "clean": "turbo run clean && rm -rf node_modules",
     "create:graph": "turbo build --graph=dependency-graph.png",
     "create:package": "yarn hygen generator create-package with-prompt"
