+++ packages/components/package.json
@@ -13,6 +13,7 @@
     "dist/**"
   ],
   "scripts": {
+    "lint": "TIMING=1 eslint --ext .ts,.tsx src/",
     "build": "tsup src/index.ts --clean --format esm,cjs --dts --minify --external react",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
   },
