+++ packages/components/src/Button/Button.tsx
@@ -14,4 +14,6 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) => {
   );
 });
 
+Button.displayName = 'Button';
+
 export default Button;
