+++ packages/css/src/system-props/color.ts
@@ -1,4 +1,3 @@
-import * as CSS from 'csstype';
 import { scales as s } from './scales';
 
 const color = {
@@ -6,20 +5,5 @@ const color = {
   textColor: s.colors('textColor'),
   borderColor: s.colors('borderColor'),
 };
-
-export interface ColorCSSProperties {
-  /**
-   * The color CSS property sets the foreground color value of an element's text and text decorations, and sets the currentcolor value.
-   */
-  color?: CSS.Property.Color | string;
-  /**
-   * The textColor CSS property sets the foreground color value of an element's text and text decorations, and sets the currentcolor value.
-   */
-  textColor?: CSS.Property.Color | string;
-  /**
-   * The border-color shorthand CSS property sets the color of an element's border.
-   */
-  borderColor?: CSS.Property.BorderColor | string;
-}
 
 export default color;
