+++ packages/css/src/system-props/index.ts
@@ -5,4 +5,3 @@ export { default as size } from './size';
 export { default as typography } from './typography';
 export { shouldTransformToVarFunc } from './scales';
 export type * from './types';
-export type { AliasesCSSProperties } from './aliases';
