+++ packages/utilities/package.json
@@ -15,6 +15,7 @@
   "scripts": {
     "test": "jest",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
+    "lint": "TIMING=1 eslint --ext .ts,.tsx src/",
     "build": "tsup src/index.ts --clean --format esm,cjs --dts --minify"
   },
   "devDependencies": {},
